Move date library to packages
