+++ packages/date/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@woocommerce/date",
+  "version": "1.0.0-alpha.0",
+  "description": "WooCommerce date utilities.",
+  "author": "Automattic",
+  "license": "GPL-2.0-or-later",
+  "keywords": [
+    "wordpress",
+    "woocommerce",
+    "date"
+  ],
+  "homepage": "https://github.com/WooCommerce/wc-admin/tree/master/packages/date/README.md",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/WooCommerce/wc-admin.git"
+  },
+  "bugs": {
+    "url": "https://github.com/WooCommerce/wc-admin/issues"
+  },
+  "main": "build/index.js",
+  "module": "build-module/index.js",
+  "react-native": "src/index",
+  "dependencies": {
+    "@babel/runtime": "^7.0.0",
+    "@wordpress/date": "2.0.3",
+    "@wordpress/i18n": "2.0.0",
+    "lodash": "^4.17.11",
+    "moment": "^2.22.2"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
